revert the last change
