+++ vm_network_migration/module_helpers/backend_service_helper.py
@@ -16,9 +16,9 @@
 from googleapiclient.http import HttpError
 from vm_network_migration.errors import *
 from vm_network_migration.modules.backend_service_modules.external_backend_service import ExternalBackendService
+from vm_network_migration.modules.backend_service_modules.internal_backend_service import InternalBackendService
 from vm_network_migration.utils import initializer
-from vm_network_migration.modules.backend_service_modules.regional_internal_backend_service import RegionalInternalBackendService
-from vm_network_migration.modules.backend_service_modules.global_internal_backend_service import GlobalInternalBackendService
+
 
 class BackendServiceHelper:
     @initializer
@@ -56,12 +56,6 @@ class BackendServiceHelper:
                                                   self.network,
                                                   self.subnetwork,
                                                   self.preserve_instance_external_ip)
-                elif self.backend_config['loadBalancingScheme'] == 'INTERNAL':
-                    return GlobalInternalBackendService(self.compute, self.project,
-                                                  self.backend_service_name,
-                                                  self.network,
-                                                  self.subnetwork,
-                                                  self.preserve_instance_external_ip)
                 else:
                     raise UnsupportedBackendService(
                         'The typeof the backend service is not supported. Migration is terminating.')
@@ -74,7 +68,7 @@ class BackendServiceHelper:
                 raise e
             else:
                 if self.backend_config['loadBalancingScheme'] == 'INTERNAL':
-                    return RegionalInternalBackendService(self.compute, self.project,
+                    return InternalBackendService(self.compute, self.project,
                                                   self.backend_service_name,
                                                   self.network,
                                                   self.subnetwork,
